srun wrapper for aprun (cray) enhancements
If the srun wrapper is executed when there is no job allocation, then
create an allocation using salloc and call the srun wrapper again so
that we can configure memory limits in aprun's execute line. Without
this change, aprun would lack the memory allocation information and
the task launch would fail if the job were allocated less than the
full node.