sinfo: Fix incorrect timestamp in sinfo -R output
sinfo does not check for differing reason timestamps when coalescing output
for sinfo -R and -Rl. This leads to incorrect timestamps in output if two
nodes are down with the same reason string but at a different time.

This fix adds a timestamp_reason_flag and a test to determine if
timestamps are equivalent before coalescing lines of output.

(Note: This patch is not an ideal fix for this issue. It would be much
better if the sinfo output handlers (_print_* in sinfo/print.c) could
also be called by sinfo to see if two lines of output were going
to be equivalent and could be coalesced. This would do away with the
separately kept "match_flags" which (as is seen here) are cumbersome
to maintain.)
Patch from Mark Grondona, LLNL.